refactor: move locale utils to a new file
